requirements.txt
azure-functions>=1.11.0
azure-storage-blob>=12.14.0
extract-msg>=0.41.0
python-dateutil>=2.8.2
